Added extra filtrations and added more structure to the files and code.
The filtrations that are added are:
- The alpha-dtm filtration (a hybrid rips version)
- The rips filtration

An important change to the way the way the data is generated is that the size of the box of
the original dataset is also added as an initialization.

The old exploration.ipynb is split into showcase.ipynb and exploration.ipynb. The former
will showcases ready to run code, while exploration.ipynb will show code that is not fully ready yet.